getRedditImages/requirements.txt
praw~=7.1.4
Pillow~=8.1.0
python-dotenv~=0.15.0
grpcio~=1.35.0
google-cloud-vision~=2.2.0
psycopg2-binary~=2.8.6